requirements.txt
openai==1.46.1
pandas==2.2.2
python-dotenv==1.0.1